+++ packages/xchain-ethereum/src/types/index.ts
@@ -1,2 +1,2 @@
-export * from './ethplorer-api-types'
+export * from './etherscan-api-types'
 export * from './client-types'